+++ src/lib.rs
@@ -7,5 +7,7 @@ pub mod legend;
 pub mod asciigraph;
 
 pub use crate::color::AnsiColor;
-pub use crate::options::{CharSet, Config, ZeroLine, Threshold, DEFAULT_CHAR_SET, create_char_set};
+pub use crate::options::{
+    CharSet, Config, ZeroLine, Threshold, DEFAULT_CHAR_SET, create_char_set
+};
 pub use crate::asciigraph::{plot, plot_many};